[Battleground] Reworked BattlegroundQueue system this should close
https://github.com/slowtorta/turtlewow-bug-tracker/issues/1612

+++ src/game/Battlegrounds/BattleGroundMgr.h
@@ -305,7 +305,7 @@ class BattleGroundMgr
 
         /* Battlegrounds */
         BattleGroundSet m_BattleGrounds[MAX_BATTLEGROUND_TYPE_ID];
-        std::vector<uint32> m_QueueUpdateScheduler;
+        std::vector<uint64> m_QueueUpdateScheduler;
         typedef std::set<uint32> ClientBattleGroundIdSet;
         ClientBattleGroundIdSet m_ClientBattleGroundIds[MAX_BATTLEGROUND_TYPE_ID][MAX_BATTLEGROUND_BRACKETS]; //the instanceids just visible for the client
         bool   m_Testing;
